add dispatcher content to dispatcher page
