+++ frontend/src/components/roles/dispatcher/formik-field.js
@@ -1,0 +1,24 @@
+import React from "react";
+import TextField from "@material-ui/core/TextField";
+import {ErrorMessage} from "formik";
+
+export default (props) => {
+    const {formikProps, formikFieldName, id, label} = props;
+    return (
+        <React.Fragment>
+            <TextField
+                margin="dense"
+                id={id}
+                label={label}
+                type="text"
+                onChange={formikProps.handleChange}
+                onBlur={formikProps.handleBlur}
+                value={formikProps.values[formikFieldName]}
+                fullWidth
+            />
+            <label style={{color: "red"}}>
+                <ErrorMessage name={formikFieldName}/>
+            </label>
+        </React.Fragment>
+    );
+};
